Merge GIO interface.
There are still things to do, mentioned in gio/TODO.  The interface is based on
whatever was available in GLib 2.20.0.  Changes after that have not been looked
at.  The code has only been lightly tested, but heavily reviewed.  I have
rewritten it three times to make it as succinct, both in C code and in future
Ruby code, as possible.  More work is needed, but this really needed to get out
the door, so here it is.